requirements.txt
flask
gunicorn
flask_cors
psycopg2-binary